splitted dgp decoder into multiple files (#55)
* splitted dgp decoder into multiple files

* fixed readme

* Fixed typing bug

* added class id mapping (#56)

* added class id mapping

* renamed test class
