+++ paralleldomain/decoding/dgp/dtos.py
